requirements.txt
jsonrpclib-pelix
lxml
newspaper3k
nltk
numpy
pika
pymongo
python-dateutil
redis
requests
scipy
sklearn
tensorflow
watchdog